requirements.txt
google-cloud-firestore==0.31.0
futures==3.2.0; python_version == "3"